Feat: improved transfer function branch splitting (#104)
* feat: added cos image planes

Preferentially samples at approximately 1/3 and 2/3 of the domain. Could
be useful for some edge cases.

* feat: improved branch split algorithm

Split branches off of pre-determined minima and maxima, therefore able
to allocate the branches in full and decide on which is the upper and
lower after they have been assembled.

The sorted interpolation now also first sets gstar extrema to 0.0 and
1.0 to avoid any erroneous extrapolation.

Toyed around with a few alternative methods for finding the extremal
redshifts when calculating the transfer functions. Some loose end code
in this commit solves using Optim for the extrema, however this is
largely untested and therefore not yet used.

* chore: bump version