Fix: more robust transfer functions (#107)
* fix: make tracing ad-compatible

Fixed some unintended type tightening in various constraint utility
functions.

* fix: more robust transfer functions

Transfer function calculations now use AD to determine the Jacobian,
which proves to be much more stable (and faster) than the stenciling
method previously used.

Added a variety of warning messages to help determine when transfer
function branches may have numerical problems.

The domain of gstar is now trimmed at the edges to avoid values where
terms like

        g * ( 1 - g )

reduce the numerics to zero. The endpoints between the branches are now
interpolated to form closed loops.

Improve theta sampling when calculating the g and J arrays to oversample
where extrema are expected, improving general convergence even at
comparatively low N.

Removed the cubic interpolation code temporarily, to prefer using
`extrema`, since the new theta domain ensures much better numerical
estimates and can often break the interpolation. The interpolation seems
to return NaN values when used in conjunction with new sampling.

Slowed the rate at which the observer is moved backwards when
calculating a batch of transfer functions to improve root finding.

* test: updated cunningham transfer function values

* chore: bump version

* fix: modify redshift for kerr metric

Numerically identical, but slightly cleaner implementation.

* fix: fortify transfer functions a bit more

Updated the precision solver and transfer function stepper to work for
the full domain, bar a few warnings. Specifically, for `a=0.0`, the
integrator seems to struggle precisely when calculating the Jacobian via
AD. The numerical result is entirely fine, but the console is bombarded
with warnings.

* style: format tests

* fix: offset angles on image plane

The coordinate singularity was causing the Jacobian calculations to fail
with the AD backend. Offset by a constant amount to fix this.

* test: update transfer function values