feat: use generated functions for geodesic equation
Results in a 6x speedup in evaluating the geodesic equation once the
Jacobian and inverse metric has been determined. Lends itself to
slightly better generics when other metric classes are fully
implemented.

Uses Julia generated functions to precompute the functional form of the
matrices, and then uses matrix multiplication to quickly and efficiently
evaluate the static arrays.

Overall performance boost of a render is approx 10%.