feat: optical properties
Added the ability to give traits to accretion geometry structs that
specify the optical properties, which are currently limited to optically
thick (opaque) and optically thin (translucent). These are then used if
radiative transfer traces are being calculated to determine which
geometry terminates the integration, and which should toggle the within
geometry state.

bug: composite geometry with overlapping regions for radiative transfer
seems to get stuck in a callback loop close to the event horizon.