feat: update coronal models
Coronal models should now implement the `sample_position_velocity`
function, which returns pairs of static vectors describing the local
position and velocity of a portion of the source.

broken: `source_velocity` has been removed, which breaks some of the
emissivity and disc flux calculations. With a new anticipated API for
emissivity profiles, the source velocity can be stored along with the
geodesic point to amend this.