feat: semi-analytic 2d transfer functions
Using the Cunningham transfer functions with timing information to
integrate the 2d transfer functions and bin according to the time delay.

todo: still a few little bugs in interpolating the gstar-time curves,
which is causing some of the time bins to be completely off. Will need a
similar prescription as with the flux edge fixes to account for this.