chore: refactor code and disc profile structures
Seperated the disc profile structures into their own logical units, so
that all relevant code to a particular structure (e.g. Voronoi) is in
the same place.

Added a new dispatch to `emissivity_profile` that raises a descriptive
error if an attempt to calculate a profile without a sampler and valid
implementation is invoked.

Rename `sky-geometry.jl` to `samplers.jl`, as it better illustrates what
that file does.