feat: spacetime concious integration parameters
The integration parameters now hold the metric for the purpose of both
dispatching and unpacking geodesic problems and solutions. This has the
benefit that the integration function can now be pure, without any
closure captures, which hopefully should aid the GPU implementation.

Also conveys a slight performance improvement of around 10%.