fix: radiative transfer formulation
Whilst trying to recreate plots from Gold+20, the current formulation of
radiative transfer produced incorrect figures. Modifying the covariant
formulation as in Eq. (2) seems to agree with the figures well.

Seperated the interface to calculate velocity and abs/em coefficients
seperately, such that the model receives the frequency in the local
frame, and not the observed frequency.

todo: check if these changes are correct?