feat: use datum plane when root-finding offsets
Since part of the disc may be obscured, we instead root find a radius on
a datum plane at the height of the disc. This is **not a perfect**
solution, since it technically only works for well-defined cross section
functions. Furthermore, obscuration must then be later recalculated.

This does however allow the minima and maxima of the redshift to be
determined for a given ring, and therefore for transfer functions to
later be calculated for thick discs.