feat: data accumulator for transfer functions
Instead of having to pass around a bunch of arrays, this commit changes
the transfer function calculations to use a utility accumulator struct
with a handful of methods. Also simplifies the API and paves the way to
obscured transfer functions nicely.

Added a utils file to `src/transferfunctions` to keep verbose utility
methods in.