feat: nan interpolator and utility functions
NaN linear interpolator will handle NaNs subtly differently from the
DataInterpolations library, in that if the knot closest to the current
point is NaN, it will return a default value (zero) instead.

The NaN interpolator is unused in this commit, but kept for possible
future use?

Added `_equitorial_project` and `_zero_if_nan` utility functions to make
those operations consistent in the codebase.