feat: progress on thick disc tf integration
Largely untested but results are starting to make sense. Defer checking
for visibility and just use NaNs as place holders, and then get the
integration routines to handle the NaNs.

note: in `src/line-profiles.jl` it is important that the max radius on
the integration plane is sensible for the binning method to work.