feat: added _gtol_error function
Better way to handle the uncertainty scaling in the geometry error. This
part of the code seems to be quite temperamental, and makes a surprising
difference to the transfer functions at distant radii.

This function is also extremely fiddly, so I've just used an ad-hoc
radial dependence that makes all the tests pass, but there is much room
for improvement.