feat: make_circular_velocity function
Some metrics may require computing the angular velocity outside of the
equatorial plane, which should be the job of the routines interpolating
the redshift. This was however not the case.

This commit adds a functor that returns a velocity function, and allows
metrics to override the velocity function to facilitate different
behaviours. This has been done for the NoZMetric, with some limitations.