feat: use SimpleNonlinearSolve for offset solving
Changed from Roots.jl to SimpleNonlinearSolve.jl to get the
time-per-solve down from about 20ms to 2.7ms on well defined problems.
An issue that persists is that for small radii on the disc, the solver
has a much harder time converging, and sometimes converges badly.

This commit attempts to make things a bit more robust, by adding a
`best` parameter to track the best solution so far incase the one
returned by the root finder projects poorly, and by trying to keep the
manifold over which we optimize as smooth as possible. This means, for
instance, determining the difference from the projected radius
irrespective of disc intersection.

The biggest win of this commit however is that calculating a set of
interpolated transfer functions has gone from ~20 seconds to about 5, so
we get a 4x speedup, with barely any (if any) loss in precision.