fix: ignore root solver errors on thick discs
When trying to determine if a patch of a thick disc is visible, we run
through the root solver again to determine whether the point on the
datum plane and the point on the thick accretion disc are within some
distance of eachother.

It is not guarunteed that the root solver will be able to find the
correct point for some emission radius when tracing over the thick disc.
This is because that point might be obscured, which is effectively the
whole point of re-tracing. So we wrap the root solving attempt with a
try-catch, ignore any error that might occur, and just set the radius to
NaN so that the point is later classed as not visible.