feat!: geodesic points store their integ params
GeodesicPoint now keeps its integration parameters instead of just the
status codes so that it is easier for downstream users to modify the
integration parameters and pull out the results.

There was a bug where EnsembleEndpointThreads reuses the same allocated
status after unpacking the solution, so to avoid overwriting the result,
we have to create a deepcopy of the parameters. This isn't really ideal,
and maybe there should be an unpack routine introduces, so some struct
of arrays solution to unpack everything into nice contiguous memory.