feat: photon rings trace
Can now tracing the number of windings around a given inclined plane to
do things like calculating lensing bands or photon rings.

The number of windings returned is the integral number of crossings of
the plane. That means, to be consistent with the literature, a winding
of 2 is a n=1 photon ring, winding of 4 is n=2, and so on.