feat: 5x speedup in transfer func integration
Another classic example of where using standard algorithms communicates
intent better, both to the reader and the compiler. Seems the quadrature
sum now gets unfolded and SIMD'd in a way that would have been tedious
to write by hand.

Results in almost a linear speedup with the number of quadrature points.