feat: add VelFunc field to extended coronal models
The velocity function is now a field that can be set in the representing
struct. Implemented both the co-rotating and stationary case.

Follow-up: at some point need to check if it's a better idea to have
this as a field in the struct or just as a parameter of the struct.
Currently it's both...