feat: corona arms
Added the rest of the implementation to calculate the
point-approximation arms. This seems to work fine now for all parameter
values of interest. The only caution is that it seems to not work as
well as the old method did for stationary rings, but I am not fully sure
why as I have not yet investigated this.

Some of the boilerplate to tie this into the `emissivity_profile` API is
still missing.