feat(emissivity): better bracket estimate
For optimising the extremal radii for the point-approximation of a
ring-corona, the bracketting algorithm had previously been to find the
best estimate, and then take some dθ either side. This neglects the fact
that e.g. the maximal radius will have a neighbour that did not hit the
disc at all, in which case the maximal θ may not lie in `θ + dθ`.

This bracketing algorithm looks if the neighbour did not hit the disc,
and moves the bracket limits to their angles respectively. If the
underlying function were already unimodal, this should maintain that
property.