fix(emissivity): sort by angle, not radius
When there is a co-rotating velocity and some degree of boosting, the
underlying assumption that for any given slice, there are exactly two
geodesics that hit a given radius is not generally true. In particular,
for the case

    m = KerrMetric(1.0, 0.9982)
    d = ThinDisc(0.0, Inf)
    model = RingCorona(Gradus.SourceVelocities.co_rotating, 30.037, 1.113)

the arm calculated with

    Gradus._ring_arm!(cache, m, d, 0.63)

will not be a star-shaped region on the accretion disc, and will have
some radii that are hit multiple times.

Neglecting this by sorting by radius on the disc introduced a
bifurcation, where emissivity values from one branch of the other were
selected. This bifurcation is artificial, and sorting by θ instead
removes it.

A new problem that then is introduced is that the interpolation may now
interpolate from small r and big r, and average the entire arm across a
large number of radii. To avoid this, we cycle the sorting index to make
sure the maximal (minimal) element is last (first) in the permutation.

This seems to remove spurious bumps that previously plagued some of
these calculations.