fix(emissivity): unique radii
Turns out it is vitally crucial to remove the duplicate radii, else it
ruins calculations later down the line. I don't know of a nice way to do
this using `sortperm`, `unique!`, and `permute!`, but it doesn't seem to
be that performance critical of a path of the code, at least when
profiling for runtime. Should at some point check the memory usage.