feat: better(ish) beta interpolation slice
One of the problems with the fixed theta way of interpolating a new beta
slice is that the distribution of angles may not evenly sample the
distribution of radii on the disc. In particular, for some coronal
sources this would lead to points around the ISCO being poorly sampled,
and producing erroneous bumps or curling in the time-averaged emissivity
profile at low r.

By concatenating the two theta arrays, we can be certain that the
sampling will still cover critical regions well.

This does however, to my dismay, reintroduce the noisy emissivity
profiles that I'd set out to fix in the first place. It also means we
can't pre-allocate the slice and reuse it during interpolation, but the
performance overhead does not seem to be too impactful.