feat(emissivity): use AD for drdθ estimates
Instead of trying to numerically approximate the Δr and Δθ terms, we can
do them both in one go by using AD to calculate what the derivative of
the r projection is as a function of θ. This seems to be much more
numerically stable than what we were previously doing, and somewhat
cleans up the resulting emissivities, but the tooth-like shape is still
present.

I am a little concerned that this approach doesn't really offer that
much, but massively slows down the integration.