perf: reuse integrator for derivatives
It occurred to me that just reaching under the first layer of
abstraction in ForwardDiff is enough to setup a reusable integrator for
evaluating derivatives. This speeds up the emissivity calculations by a
factor of about 4, so now it's only twice as slow with the AD calls.
This is to be expected, as it has to do twice as many integrations.

There is more room for optimisation here, and a similar trick should be
used for the Cunningham transfer function evaluation with the Jacobians.