feat: adaptive sky emissivity methods for corona
These are not the final APIs, but the bulk of the operations.

The binning methods bin up the values calculated using an interpolation
scheme, weighted by the solid angle of that particular grid cell. The
total solid angle for each bin is also calculated, and used to normalise
the output.

When experimenting with this I'd previously used the mean in a given
cell, but there was a frustrating dependence on the ϕ grid resolution
which changed the result. Now, there is only a lower-bound on the
resolution, which is to be expected as an approximate method.

A method for interpolating the resulting emissivities binned by r and ϕ
is provided, which does a sufficiently good job provided the space is
not too sparsely sampled.

Also adds a function for approximately filling in the adaptive sky with
a dense evaluation.