feat: almost working adaptive emissivity
It occurred to be it makes much more sense to interpolate the inverse of
the emissivity, and then invert it at the end before summing. The reason
for this is that both the error scales slower in the interpolation
scheme, but also it averages out the Jacobian _before_ the inverse is
taken. This avoids the caustic that I'd been observing over-contributing
to the total emissivity profile.

There is still a bit of sensitivity to the grid resolutions, which
suggests to me there is still some weighting factor missing, but it's a
marked improvement over previous iterations.

It makes me think maybe working with inv(ϵ) might work for the
interpolated branches as well. Something to try in the future at any
rate.