feat: full adaptive refinement strategy
Uses a block-based strategy to refine the illumination profile of a
corona on an accretion disc. Splits the r, ϕ plane of the disc into a
number of blocks, evaluates some metric for each, and refines those
blocks with the lowest score until some threshold is met.

Solves most configurations of the corona to pretty high resolution in
around half a minute on my (unplugged, i.e. no turbo) laptop.

The strategy is still not perfect. In particular, for low corona but at
distant radii (x > 30, h < 2), it fails to get the very small `r`
behaviour right. Other than that, it seems to do everything else well.

Seems to relatively consistently end up evaluating around half a million
geodesics, which is not too bad (though also not ideal).