feat: image offset hinting
Try using the previous runs to hint to the root finder where the
solution for intersecting a given radius on the image plane should be.
The hint is then used as the initial solution for the next call to
`workhorse`. Adds a new field to track all of the hints from previous
interations.

This is experimental.