feat: handle convergence in emissivity differently
Broke up some of the adaptive sky tracing methods so that they can be
called individually -- e.g., finding which cells need to be refined, and
then refining them can now be easily done from other functions.

This then allows the limit check to be moved to a parent function, i.e.
`step_block!` in the adaptive solve. Step block now checks the limits
and returns an enum value as a status code, namely limit exceed,
converged, or not converged. The calling context, `adaptive_solve!`,
then handles the return code.

The new behaviour is to break off the iterative refinement if the limit
is exceeded, and I'll investigate in the future how to fix those cases.