feat: new offset_for_radius algorithm
This is the first iteration towards trying to implement a robust and
accurate root-finder for finding intersections on the accretion disc.

Similar to other recent changes in the codebase (e.g. the adaptive
solver), this new approach uses the fact that we can calculate
derivatives efficiently with the reusable-integrators when configured
with custom ForwardDiff.jl tags.

This commit implements a new root-finding algorithm based on the
Newton-Raphson algorithm.

It was my observation that the `r_img -> r_disc` map is generally
hyperbolic until `r_disc == r_EH`, which are not the easiest roots to
solve, as linear interpolations will miss the true root and hit the
tarpit. This first approach seems to perform okay.