feat: contrapoint bias in root-finder
Help the root-finder out of the tarpit by tracking a contrapoint, i.e.
the best "worse" guess. This can be used to do something akin to
bisection, except it is now weighted with a bias term that means we
don't quite bisect, but pick a new point closer to the contrapoint by
the bias fraction.

This seems to help the root-finder solve on datum planes.