feat: cycle detection and bracketing fallback
For the new root-finder, start tracking how much better/worse recent
steps were. If a cycle is detected (as may happen with Newton-Raphson
solvers), then we are likely close and we fallback to the Bisection
algorithm of Roots.jl to finish off finding the root. The bracketing
interval is given as the current guess and the contrapoint.