test: update regression tests for new root-solver
I plotted most of them to make sure they hadn't changed. The main thing
that altered was that around the extrema of the CTF, the values are
undefined, and those did change. That is largely inconsequential, since
they are later masked out of integration, but for the regression tests
they lead to subtly different values.