fix: temporary threadid fix
This should be ammended with the use of `@spawn` and tasks in the
future, but I want to merge the `thesis` branch before addressing that.

Julia 1.12 has interactive threads, so the thread IDs no longer behave
as expected. This was always an anti-pattern I suppose, so should
switch to the better semantics.