test: update regression values
There have been a few definition changes that have affected some of the
regression test values, such as changes to how the geometric grid is
defined.

The changes is flux in the 2D transfer functions is slightly more
alarming, but I've checked that the documentation examples can be
reproduced, as well as the figures from the paper, so I think it might
just be some noise artifact. Those tests should probably be updated.

Also removed the Voronoi tessellation tests, since that part of the code
is no longer used and the underlying library keeps changing its
primitives.

All dependencies on that library should be removed in the future.