fix: thick disc transfer functions
The new algorithm is extremely sensitive to having the right beta_0
offset parameter, and if it's too big, will take along time to fail,
spurting "exceeded maxiters" warnings. That is obviously not ideal and
should be examined in more detail in the future, but my priority at the
moment is to get these commits merged in.

Another aspect that was driving the CTF solving time up was running the
root-solver twice, once over the datum plane and again against the
actual thick disc. This commit removes that, and just checks if the
impact parameters from the datum plane actually hit close enough on the
thick disc.

Updated the regression test values.